GHV5/requirements.txt
numpy>=1.24
pyserial>=3.5
scikit-learn>=1.3
pandas>=1.5
pygame>=2.5
